requirements.txt
asgiref==3.7.2
boto3==1.28.21
botocore==1.31.21
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
celery==5.2.7
cffi==1.16.0
charset-normalizer==3.2.0
cloudinary==1.33.0
cryptography==41.0.7
Django==4.2.3
django-celery-beat==2.5.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-filter==23.2
django-rest-framework==0.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
fcm-django==2.0.0
firebase-admin==6.3.0
flower==1.2.0
google-api-core==2.15.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-cloud-core==2.4.1
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
msgpack==1.0.7
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.8
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
redis==4.3.4
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
